feat: added title based on routes
