+++ apps/frontend/src/app/(app)/inbox/page.tsx
@@ -1,4 +1,13 @@
+import { Metadata } from "next"
+
 import { InboxPage } from "@/src/components/Inbox/InboxPage"
+import generateMetadata from "@/src/utils/seo"
+
+export const metadata: Metadata = generateMetadata({
+  path: "/inbox",
+  title: "Inbox",
+  description: "engineered for makers",
+})
 
 const Inbox: React.FC = () => {
   return (
