+++ apps/frontend/src/app/(app)/space/page.tsx
@@ -1,5 +1,15 @@
 import * as React from "react"
 
+import { Metadata } from "next"
+
+import generateMetadataHelper from "@/src/utils/seo"
+
+export const metadata: Metadata = generateMetadataHelper({
+  path: "/spaces",
+  title: "Spaces",
+  description: "engineered for makers",
+})
+
 const SpacePage: React.FC = () => {
   return (
     <section className="size-full overflow-auto bg-background px-8 py-16">
