+++ apps/frontend/src/app/(app)/this-week/page.tsx
@@ -1,6 +1,15 @@
 import React from "react"
 
+import { Metadata } from "next"
+
 import { ThisWeekPage } from "@/src/components/ThisWeek/ThisWeekPage"
+import generateMetadataHelper from "@/src/utils/seo"
+
+export const metadata: Metadata = generateMetadataHelper({
+  path: "/this-week",
+  title: "This Week",
+  description: "engineered for makers",
+})
 
 const ThisWeek: React.FC = () => {
   return (
