+++ apps/frontend/src/app/(app)/today/page.tsx
@@ -1,5 +1,16 @@
-import TodayPage from "@/src/components/Today/TodayPage"
+import { Metadata } from "next"
 
-export default function Today() {
+import TodayPage from "@/src/components/Today/TodayPage"
+import generateMetadataHelper from "@/src/utils/seo"
+
+export const metadata: Metadata = generateMetadataHelper({
+  path: "/today",
+  title: "Today",
+  description: "engineered for makers",
+})
+
+const Today: React.FC = () => {
   return <TodayPage />
 }
+
+export default Today
